requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
redis==5.0.4
Flask-Caching==2.3.0
gunicorn==21.2.0